refactor(frontend): split app logic into modules
