+++ frontend/constants.js
@@ -1,0 +1,9 @@
+export const DEFAULT_CONVERSATION_TITLE = '新的会话';
+export const CONVERSATION_TITLE_MAX_LENGTH = 20;
+
+export const STORAGE_KEYS = {
+  conversations: 'okc.conversations',
+  current: 'okc.conversations.current',
+};
+
+export const SERVICES = ['chat', 'image', 'speech', 'sound_effects', 'asr'];
