+++ frontend/elements.js
@@ -1,0 +1,26 @@
+export const chatMessages = document.getElementById('chat-messages');
+export const chatForm = document.getElementById('chat-form');
+export const userInput = document.getElementById('user-input');
+export const statusPill = document.getElementById('status-pill');
+export const modelLogList = document.getElementById('model-log');
+export const modelLogEmpty = document.getElementById('model-log-empty');
+export const webPreviewFrame = document.getElementById('web-preview-frame');
+export const webPreviewEmpty = document.getElementById('web-preview-empty');
+export const openWebPreviewButton = document.getElementById('open-web-preview');
+export const pptPreviewContainer = document.getElementById('ppt-preview');
+export const pptPreviewEmpty = document.getElementById('ppt-preview-empty');
+export const togglePptModeButton = document.getElementById('toggle-ppt-mode');
+export const modelLogTemplate = document.getElementById('model-log-item-template');
+export const pptSlideTemplate = document.getElementById('ppt-slide-template');
+export const configForm = document.getElementById('config-form');
+export const configStatus = document.getElementById('config-status');
+export const settingsToggle = document.getElementById('settings-toggle');
+export const settingsOverlay = document.getElementById('settings-overlay');
+export const settingsDrawer = settingsOverlay?.querySelector('.settings-drawer') ?? null;
+export const settingsCloseButtons = document.querySelectorAll('[data-action="close-settings"]');
+export const historySidebar = document.getElementById('history-sidebar');
+export const historyToggle = document.getElementById('history-toggle');
+export const historyPanel = document.getElementById('history-panel');
+export const conversationList = document.getElementById('conversation-list');
+export const conversationEmptyState = document.getElementById('conversation-empty');
+export const newConversationButton = document.getElementById('new-conversation');
