+++ frontend/utils.js
@@ -1,0 +1,33 @@
+export function generateId() {
+  if (typeof crypto !== 'undefined' && crypto.randomUUID) {
+    return crypto.randomUUID();
+  }
+  return `id-${Date.now().toString(16)}-${Math.random().toString(16).slice(2, 10)}`;
+}
+
+export function formatConversationTime(isoString) {
+  if (!isoString) return '';
+  const date = new Date(isoString);
+  if (Number.isNaN(date.getTime())) return '';
+  return date.toLocaleString('zh-CN', {
+    month: '2-digit',
+    day: '2-digit',
+    hour: '2-digit',
+    minute: '2-digit',
+  });
+}
+
+export async function fetchJson(url, options = {}) {
+  const response = await fetch(url, options);
+  if (!response.ok) {
+    let detail = '';
+    try {
+      const body = await response.json();
+      detail = body?.detail || body?.message || '';
+    } catch (error) {
+      // ignore json parse errors
+    }
+    throw new Error(detail || `请求失败：${response.status}`);
+  }
+  return response.json();
+}
